feat: refactor user profile retrieval into separate file and restore functionality
